frontend: Protect AddChirpPage
Add ProtectedRoute component that prevents a route from being
accessible by unauthenticated users. Users will be redirected with
Auth0 upon accessing that route. After logging in, they will be
returned to that route.

Add Spinner component that renders a huge spinner in order to indicate
to the user that they are being redirected to the login.