Implement proper redirect-after-login
Implement a proper redirect after logging in from directly
going to a ProtectedRoute by implementing a custom Auth0Provider
with a custom onRedirectCallback.

Add configuration for the nginx docker image to always redirect
everything to /index.html so that the BrowserRouter works.
Add the same config to public/_redirects for Netlify.