Fix recreation of cache
Because React-Router v6's useNavigate changes on every update to the route,
the Apollo provider was being re-rendered on every page change, which
re-created the cache each time. Put the creation of the cache outside
of the component now in order to keep it around.

Still not great that RR is doing this, keep an eye on this issue:
https://github.com/remix-run/react-router/issues/7634